refactor: Include isAuthenticated flag in shop routes
Include the isAuthenticated flag in the shop routes of the shop.controller.js file. This flag is used to determine if the user is authenticated and logged in. By including this flag in the rendered views, the application can display different content based on the user's authentication status. This commit improves the security and user experience of the shop functionality.