feat: Add forgot password link to login page
A link to reset the password has been added to the login page. This enhancement improves the user experience by providing an easy way for users to reset their passwords if they have forgotten them. The link redirects to the reset password page. This change complements the recent commits that added the reset password functionality and updated the email sending capabilities.