refactor: Add 500 error page
The code changes in the `500.ejs` file add a new error page for handling internal server errors. This enhancement ensures that a proper response with a status code of 500 is sent to the client in case of an internal server error. This improves the reliability and user experience of the application.