feat: Add view more button to home page
The code changes in main.css and product-list.ejs introduce a new "view more" button to the home page. This button allows users to load additional products without refreshing the page. The button is centered and styled with a fit-content width and a margin-top of 3rem. This feature enhancement improves the usability and user experience of the application.