feat: Add image URL to post in addPost route
This commit modifies the addPost route in feed.controller.js to include the image URL of the uploaded image in the post object. The image URL is obtained from the req.file.path and is properly formatted by replacing backslashes with forward slashes. This enhancement improves the functionality of the backend application by associating the uploaded image with the created post.