feat: Update Feed.js to support editing and deleting posts
This commit modifies the `Feed.js` file to support editing and deleting posts. The `editPost` functionality is added to the `POST` request, allowing users to update a post by its ID. The `deletePostHandler` function is also updated to send a `DELETE` request to the server, deleting a post by its ID. These changes enhance the functionality of the frontend application by providing the ability to edit and delete posts.