feat: Add signup functionality to auth controller
This commit adds the signup functionality to the `auth.controller.js` file. The `signup` function handles the creation of a new user by validating the request body, hashing the password, and saving the user to the database. This feature enhances the functionality of the backend application by providing the ability to sign up new users.