feat: Update post model to use ObjectId for creator field
This commit modifies the `post.model.js` file to update the `creator` field in the `PostSchema` to use `Schema.Types.ObjectId` and reference the `User` model. This change ensures that the `creator` field correctly stores the ID of the user who created the post, improving the data integrity and functionality of the backend application.