feat: Populate creator field in getPosts API response
This commit modifies the `feed.controller.js` file to populate the `creator` field in the response of the `getPosts` API. By using the `populate` method, the creator information is fetched from the database and included in the API response. This enhancement improves the completeness of the post data and provides additional information about the post creator.