feat: Add authorization check to addPost resolver
This code change adds an authorization check to the addPost resolver in the resolvers.js file. If the user is not authenticated (req.isAuth is false), an error with a 401 Unauthorized status code is thrown. This enhancement improves the security of the application by ensuring that only authenticated users can add posts.